Fixed jump -> now nice and smooth + changed TILE from int to float. Need to add some comments

+++ 1415_2TIB4_aventure/animations.cpp
@@ -1,7 +1,7 @@
 #include "flib.h"
 #include "animations.h"
 
-const int TILE = 32;		// Tiles size
+const float TILE = 32;		// Tiles size
 
 const int XSCREENLENGHT = 15;	// Number of tiles on x
